+++ configurations/csbc-pson/synapseConfigs/index.ts
