+++ configurations/csbc-pson/synapseConfigs/publications.ts
