+++ configurations/csbc-pson/synapseConfigs/studies.ts
