+++ configurations/nf/headerConfig.ts
@@ -2,7 +2,6 @@ import { HomePageHeaderConfig } from '../types/portal-config'
 
 const homePageHeader: HomePageHeaderConfig = {
   title: 'A home for Neurofibromatosis research resources',
-  // tslint:disable-next-line:max-line-length
   summary: 'The NF Data Portal is a conduit for ongoing and published research activity related to Neurofibromatosis and Schwannomatosis. It contains data, analysis tools, and publications with the goal of accelerating the development of new treatments for NF.',
 }
