+++ configurations/nf/synapseConfigs/publications.ts
