+++ configurations/nf/synapseConfigs/studies.ts
